// packages/gitea-client/generated/gitea.d.ts
export interface components {
    schemas: {
// ...
        Repository: {
            allow_fast_forward_only_merge?: boolean;
            allow_manual_merge?: boolean;
            allow_merge_commits?: boolean;
            allow_rebase?: boolean;
            allow_rebase_explicit?: boolean;
            allow_rebase_update?: boolean;
            allow_squash_merge?: boolean;
            archived?: boolean;
            /** Format: date-time */
            archived_at?: string;
            autodetect_manual_merge?: boolean;
            avatar_url?: string;
            clone_url?: string;
            /** Format: date-time */
            created_at?: string;
            default_allow_maintainer_edit?: boolean;
            default_branch?: string;
            default_delete_branch_after_merge?: boolean;
            default_merge_style?: string;
            description?: string;
            empty?: boolean;
            external_tracker?: components["schemas"]["ExternalTracker"];
            external_wiki?: components["schemas"]["ExternalWiki"];
            fork?: boolean;
            /** Format: int64 */
            forks_count?: number;
            full_name?: string;
            has_actions?: boolean;
            has_code?: boolean;
            has_issues?: boolean;
            has_packages?: boolean;
            has_projects?: boolean;
            has_pull_requests?: boolean;
            has_releases?: boolean;
            has_wiki?: boolean;
            html_url?: string;
            /** Format: int64 */
            id?: number;
            ignore_whitespace_conflicts?: boolean;
            internal?: boolean;
            internal_tracker?: components["schemas"]["InternalTracker"];
            language?: string;
            languages_url?: string;
            licenses?: string[];
            link?: string;
            mirror?: boolean;
            mirror_interval?: string;
            /** Format: date-time */
            mirror_updated?: string;
            name?: string;
            /**
             * @description ObjectFormatName of the underlying git repository
             * @enum {string}
             */
            object_format_name?: "sha1" | "sha256";
            /** Format: int64 */
            open_issues_count?: number;
            /** Format: int64 */
            open_pr_counter?: number;
            original_url?: string;
            owner?: components["schemas"]["User"];
            parent?: components["schemas"]["Repository"];
            permissions?: components["schemas"]["Permission"];
            private?: boolean;
            projects_mode?: string;
            /** Format: int64 */
            release_counter?: number;
            repo_transfer?: components["schemas"]["RepoTransfer"];
            /** Format: int64 */
            size?: number;
            ssh_url?: string;
            /** Format: int64 */
            stars_count?: number;
            template?: boolean;
            topics?: string[];
            /** Format: date-time */
            updated_at?: string;
            url?: string;
            /** Format: int64 */
            watchers_count?: number;
            website?: string;
        };
// ...
        Tag: {
            commit?: components["schemas"]["CommitMeta"];
            /** @description The ID (SHA) of the tag */
            id?: string;
            /** @description The message associated with the tag */
            message?: string;
            /** @description The name of the tag */
            name?: string;
            /** @description The URL to download the tarball archive */
            tarball_url?: string;
            /** @description The URL to download the zipball archive */
            zipball_url?: string;
        };
// ...
        User: {
            /** @description Is user active */
            active?: boolean;
            /** @description URL to the user's avatar */
            avatar_url?: string;
            /** Format: date-time */
            created?: string;
            /** @description the user's description */
            description?: string;
            /** Format: email */
            email?: string;
            /**
             * Format: int64
             * @description user counts
             */
            followers_count?: number;
            /** Format: int64 */
            following_count?: number;
            /** @description the user's full name */
            full_name?: string;
            /** @description URL to the user's gitea page */
            html_url?: string;
            /**
             * Format: int64
             * @description the user's id
             */
            id?: number;
            /** @description Is the user an administrator */
            is_admin?: boolean;
            /** @description User locale */
            language?: string;
            /** Format: date-time */
            last_login?: string;
            /** @description the user's location */
            location?: string;
            /** @description login of the user, same as `username` */
            login?: string;
            /**
             * @description identifier of the user, provided by the external authenticator (if configured)
             * @default empty
             */
            login_name: string;
            /** @description Is user login prohibited */
            prohibit_login?: boolean;
            /** @description Is user restricted */
            restricted?: boolean;
            /**
             * Format: int64
             * @description The ID of the user's Authentication Source
             */
            source_id?: number;
            /** Format: int64 */
            starred_repos_count?: number;
            /** @description User visibility level option: public, limited, private */
            visibility?: string;
            /** @description the user's website */
            website?: string;
        };
// ...
    };
// ...
}